backend/auth_service/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
cffi==1.17.1
click==8.2.0
cryptography==36.0.2
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
